modified:   .trunk/trunk.yaml

+++ profiles/NICHOLAS85/bootstrap.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env zsh
-# shellcheck disable=SC2250,2312,1091
-curl "https://sh.rustup.rs" -sSf | sh -s -- -y
-echo "source ${HOME}/.cargo/env" >>"${HOME}"/.zshenv
 
-source "$HOME/.cargo/env"
+command curl "https://sh.rustup.rs" -sSf | sh -s -- -y || true
 
-sudo apt update
-sudo DEBIAN_FRONTEND=noninteractive apt install --yes tlp lscolors fd-find fzf
-cargo install lsd
+if [[ -f "${HOME}/.cargo/env" ]]; then
+  print "source ${HOME}/.cargo/env" >>"${HOME}"/.zshenv
+  source "${HOME}/.cargo/env"
+
+  sudo apt update
+  sudo DEBIAN_FRONTEND=noninteractive apt install --yes tlp lscolors fd-find fzf
+  cargo install lsd
+fi
